Rework getFile logic
