+++ index.js
@@ -3,41 +3,16 @@ const path = require('path');
 
 global.appRoot = path.resolve(__dirname);
 
+const lorco = require('./src/lorco');
+
 const getFile = require('./helpers/getFile');
 const getLanguage = require('./helpers/getLanguage');
-const transformToRGBA = require('./helpers/transformToRGBA');
-const createVariable = require('./helpers/createVariable');
-const createFile = require('./helpers/createFile');
 
 const file = getFile();
 const language = getLanguage();
 
-const lorco = (file, language) => {
-  return ns.read(file)
-    .then(sketch => {
-      const { symbols } = sketch;
-      const fileName = path.basename(file, '.sketch');
-
-      const colors = symbols.map((symbol) => {
-        const [layer, ...others] = symbol.layers;
-        const style = layer.get('style').toJson();
-        const [fill] = style.fills;
-
-        const { color } = fill;
-
-        const rgbacolor = transformToRGBA(color);
-        const name = symbol.name;
-
-        return createVariable(name, rgbacolor, language);
-      });
-
-      createFile(`_${fileName}`, colors, language);
-
-      return colors;
-    })
-    .catch((err) => new Error('Error: ', err));
+if (file && language) {
+  return lorco(file, language);
 }
 
-lorco(file, language);
-
-module.exports = lorco;
+return console.log('No file specified');
